update heros and superpwers page

+++ superherosighting/src/main/java/com/sg/superherosighting/service/ServiceLayerImpl.java
@@ -23,6 +23,8 @@ import javax.validation.Validation;
 import javax.validation.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 
 /**
  *
@@ -54,6 +56,7 @@ public class ServiceLayerImpl implements ServiceLayer {
     private Set<ConstraintViolation<Sighting>> sightingViolations = new HashSet<>();
     private Set<ConstraintViolation<Organization>> organizationViolations = new HashSet<>();
 
+    //HERO
     @Override
     public Hero getHeroById(int id) {
 
@@ -71,7 +74,17 @@ public class ServiceLayerImpl implements ServiceLayer {
     @Override
     public Hero addHero(Hero hero) {
 
+        SuperPower superPower = new SuperPower();
+
+        superPower.setName("none");
+
+        SuperPower defaultSuperPower = getDefaultSuperPower(superPower);
+
+        hero.setSuperPower_id(defaultSuperPower.getId());
+        hero.setSuperPower(defaultSuperPower.getName());
+
         return heroDao.addHero(hero);
+
     }
 
     @Override
@@ -103,6 +116,7 @@ public class ServiceLayerImpl implements ServiceLayer {
         return superPowerDao.getSuperPowerById(id);
     }
 
+    //Super Power
     @Override
     public void insertSuperPowerHero(SuperPower superPower) {
 
@@ -117,18 +131,90 @@ public class ServiceLayerImpl implements ServiceLayer {
     @Override
     public SuperPower addSuperPower(SuperPower superPower) {
 
-        return superPowerDao.addSuperPower(superPower);
+        return getDefaultSuperPower(superPower);
+
     }
 
     @Override
-    public void updateSuperPower(SuperPower superPower) {
+    public void updateSuperPower(SuperPower superPower, BindingResult result) {
 
-        superPowerDao.updateSuperPower(superPower);
+        List<SuperPower> superPowers = superPowerDao.getSuperPowerByName(superPower.getName());
+
+        //check for existing super power with that name
+        if (superPowers.isEmpty()) {
+
+            superPowerDao.updateSuperPower(superPower);
+        } else {
+
+            //check if old name and new name is the same then send error
+            if (superPower.getName() == superPowers.get(0).getName()) {
+                FieldError error = new FieldError("superPower", "name", superPower.getName() + " is already taken.");
+                result.addError(error);
+            } else {
+                //names are diffrent update super power
+                superPowerDao.updateSuperPower(superPower);
+            }
+        }
     }
 
     @Override
     public void deleteSuperPowerById(int id) {
-        superPowerDao.deleteSuperPowerById(id);
+
+        SuperPower superPowerToDelete = superPowerDao.getSuperPowerById(id);
+
+        // create new super power
+        SuperPower superPower = new SuperPower();
+
+        superPower.setName("none");
+
+        //check if it super power called none exists
+        List<SuperPower> superPowers = superPowerDao.getSuperPowerByName(superPower.getName());
+
+        //default value doesnt exist in the db
+        if (superPowers.isEmpty()) {
+
+            //create one
+            superPower = superPowerDao.addSuperPower(superPower);
+
+            //get all heros
+            List<Hero> heros = superPowerDao.getSuperPowerDetails(id);
+
+            //add to super Power object
+            superPower.setHeros(heros);
+
+            //update all heros with id so they can be deleted
+            superPowerDao.updateSuperPower(superPower);
+
+            //delete id that was meant to be deleted
+            superPowerDao.deleteSuperPowerById(id);
+
+        } else {
+
+            //if it does exist
+            //check if power to be deleted == none , delete immediately
+            if (superPowerToDelete.getName().equals("none")) {
+
+            } else {
+                //power that will be deleted is not "none"
+
+                //get exusting one
+                superPower = superPowers.get(0);
+
+                //get all heros
+                List<Hero> heros = superPowerDao.getSuperPowerDetails(id);
+
+                //add to super Power object
+                superPower.setHeros(heros);
+
+                //update all heros with id so they can be deleted
+                superPowerDao.updateSuperPower(superPower);
+
+                //delete id that was meant to be deleted
+                superPowerDao.deleteSuperPowerById(id);
+            }
+
+        }
+
     }
 
     @Override
@@ -239,7 +325,7 @@ public class ServiceLayerImpl implements ServiceLayer {
     }
 
     @Override
-    public Hero getSuperPowerDetails(int superPowerId) {
+    public List<Hero> getSuperPowerDetails(int superPowerId) {
         return superPowerDao.getSuperPowerDetails(superPowerId);
     }
 
@@ -325,6 +411,24 @@ public class ServiceLayerImpl implements ServiceLayer {
     @Override
     public Set<ConstraintViolation<Organization>> getOrganizationViolations() {
         return organizationViolations;
+    }
+
+    private SuperPower getDefaultSuperPower(SuperPower superPower) {
+
+        List<SuperPower> superPowers = superPowerDao.getSuperPowerByName(superPower.getName());
+
+        if (superPowers.isEmpty()) {
+            //create new super power if not found
+            superPower = superPowerDao.addSuperPower(superPower);
+
+        } else {
+            //return existing
+            superPower = superPowers.get(0);
+
+        }
+
+        return superPower;
+
     }
 
 }
